Add a class and a management command for importing from PopIt
Since PopIt (a data store, API and editing interface for Popolo
data) is being retired, it would be useful to have an easy migration
path for projects that used to use PopIt to move them over to using
django-popolo.

This commit introduces a management command, popolo_create_from_popit,
to import data from a PopIt JSON data export into django-popolo's
models.  This uses a PopItImporter class, which is intended to be easy
to customize by overriding its methods to deal with the common
situation that users have been storing attributes outside the core
Popolo schema in their PopIt instance.  It also should make it
possible to use this class in a migration if you override
get_popolo_model_class.