Fix handling of model classes in migrations
Some of the code in the importer didn't work propely with the
"fake" model classes you'll get when using this in a migration;
e.g. you can't use the content_object helper property created
by a GenericForeignKey.  These changes make it easier to use
the importer in a migration, and shouldn't affect it when used
standalone.